Finalized Behaviour Tree Design
Behaviour tree functions no longer contain a state parameter, all state is to be stored in the Blackboard. Some of the nodes have internal handlers that require state, those keys must now be passed in at construction (they can be updated later if the objects are statically constructed).

Added an example of the selector node to the pet behaviour.