Expanded pet entity management
Added the PetEntity class to store the state of a pet.
Added a function for applications to receive events.
Added function to PetAI callbacks to allow the application to get its custom state data from a PetHandle. This state data is intended to be stuff like fur color, eye color, etc.
Refactored main loop to iterate over all pets.
Refactored main loop to create a pet to store the behaviour tree in.